+++ src/modules/adsb-feeder/filesystem/root/opt/adsb/extras/temperature.py
@@ -9,10 +9,16 @@ import tempfile
 # Raspberry Pi or an Orange Pi Zero 3
 #
 # for Raspberry Pi the pin number passed in is the BCM number,
-# so for GPIO22 (aka pin 15 on the 26 or 40 pin header) pass in 22
+# so for GPIO4 (aka pin 7 on the 26 or 40 pin header) pass in 4
+# -- this means the DHT22 sensor is connected to
+#
+#  pin 1 - 3.3V
+#  pin 7 - GPIO4
+#  pin 9 - GND
 #
 # for Orange Pi Zero 3 the pin number passed in is the PCxx number,
 # so for PC9 (aka pin 7 on the 26 pin header) pass in 9
+
 
 class RPI:
     def __init__(self, pin):
@@ -25,6 +31,7 @@ class RPI:
         if self.sensor.staleness() < 5:
             return self.sensor.temperature()
         return None
+
 
 class OPI:
     def __init__(self, gpio, pin):
@@ -43,6 +50,7 @@ class OPI:
             self.last_temperature = result.temperature
         return None
 
+
 if __name__ == "__main__":
     # command line arg - call with PIN number (BCM for RPi, PCxx for OPi Zero 3)
     pin = None
@@ -58,14 +66,16 @@ if __name__ == "__main__":
     if model == "OrangePi Zero3":
         import OPi.GPIO as gpio
         import opi_dht22
+
         if pin == None:
             pin = 9
         sensor = OPI(gpio, pin)
     elif model.startswith("Raspberry Pi"):
         import pigpio
         import rpi_dht22
+
         if pin == None:
-            pin = 22
+            pin = 4
         sensor = RPI(pin)
     else:
         print(f"No support for temperature sensors on |{model}|")
@@ -80,7 +90,7 @@ if __name__ == "__main__":
             continue
         temperature = sensor.get_temperature()
         if temperature != None:
-            output = { "cpu": "" }
+            output = {"cpu": ""}
             try:
                 with open("/sys/class/thermal/thermal_zone0/temp", "r") as cpu:
                     output["cpu"] = f"{int(cpu.read().strip()) / 1000:.0f}"